pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Page analyzer application"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "flask>=2.0.0",
    "gunicorn>=20.0.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "validators>=0.22.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["page_analyzer"]

